requirements.txt
Flask
Flask-PyMongo
Flask-Cors
pymongo[srv]
python-dotenv
mechanicalsoup
pandas
bcrypt
